Add scalafmt rule names basing on rules_scala docs
If you define a scalafmt rule as documented in rules_scala it won't be
recognized by intellij plugin as a Scala target, as the rule names for
rules_scala is hard-coded.

In effect the external dependencies for such targets are not discovered by
IntelliJ.

The reason is that the Scala targets generated by rules_scala do not expose
a jdeps interface, hence intellij BlazeJavaWorkspaceImporter cannot find
the dependencies for them (assuming they are not in the working set).

Docs for rules_scala scalafmt support:
https://github.com/bazelbuild/rules_scala/blob/master/docs/phase_scalafmt.md